requirements.txt
azure-storage-blob==1.5.0
python-dotenv==0.15.0
pylint==2.11.1